feat: AI 추천 종목 컴포넌트 분리 및 사용자 성향 기반 추천 기능 구현
- Top5 추천 종목 UI를 별도 컴포넌트(Top5Recommendations.jsx)로 분리하여 코드 구조 개선
- 사용자 맞춤 추천 종목 컴포넌트(UserRecommendations.jsx) 생성
  - predicted_label이 3인 종목 중 상위 5개만 필터링하여 카드 UI로 렌더링
  - 종목 클릭 시 요약 정보, 리스크 분석, 상승 확률 높은 뉴스 TOP3 표시
- 사용자 투자 성향(investment_style)을 페이지 상단에 표시
  - 예: (공격형 투자자), (중립형 투자자) 등
- Redux에 recommendUserSlice 추가 및 API 연동 완료
  - 백엔드에서 투자 성향 기반 final_score 정렬 API 사용
- AIRecommPage에서 메뉴 선택(Tab UI)에 따라 Top5 또는 사용자 추천 컴포넌트 조건 렌더링 처리
