+++ frontend/fivo-client/src/app/store.js
@@ -1,35 +1,40 @@
-// 📄 src/app/store.js
-import { configureStore } from '@reduxjs/toolkit'
-import authReducer from '../slices/authSlice'
-import storage from 'redux-persist/lib/storage'
-import { persistReducer, persistStore } from 'redux-persist'
-import { combineReducers } from 'redux'
-import { thunk } from 'redux-thunk' // ✅ 여기 수정
+import { configureStore, combineReducers } from "@reduxjs/toolkit";
+import storage from "redux-persist/lib/storage";
+import { persistReducer, persistStore } from "redux-persist";
+import { thunk } from "redux-thunk"; // ✅ thunk middleware
+import authReducer from "../slices/authSlice";
 import themeReducer from "../slices/themeSlice";
 import recommendRandomReducer from "../slices/recommendRandom5Slice";
-import favoritesReducer from '../slices/favoriteSlice'
+import recommendUserReducer from "../slices/recommendUserSlice";
+import favoritesReducer from "../slices/favoriteSlice";
 
+// ✅ persist 설정
 const persistConfig = {
-  key: 'root',
+  key: "root",
   storage,
-  whitelist: ['auth'], // ✅ auth slice만 localStorage에 저장
-}
+  whitelist: ["auth"], // 🔒 auth만 저장 (JWT 등)
+};
 
+// ✅ root reducer 구성
 const rootReducer = combineReducers({
   auth: authReducer,
   theme: themeReducer,
   recommend5: recommendRandomReducer,
+  userRecommendedStocks: recommendUserReducer,
   favorites: favoritesReducer,
-})
+});
 
-const persistedReducer = persistReducer(persistConfig, rootReducer)
+// ✅ persist 적용된 reducer 생성
+const persistedReducer = persistReducer(persistConfig, rootReducer);
 
+// ✅ store 생성
 export const store = configureStore({
   reducer: persistedReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
-      serializableCheck: false, // 🔥 redux-persist에서 필요
+      serializableCheck: false, // redux-persist의 비직렬화 경고 무시
     }).concat(thunk),
-})
+});
 
-export const persistor = persistStore(store)
+// ✅ persistor 생성
+export const persistor = persistStore(store);
