+++ frontend/fivo-client/src/main.jsx
@@ -2,18 +2,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import { Provider } from 'react-redux';
-import { store } from './app/store';
-import { GoogleOAuthProvider } from '@react-oauth/google';
-
 import App from './App';
 
+import { Provider } from 'react-redux';
+import { store, persistor } from './app/store';
+import { PersistGate } from 'redux-persist/integration/react';
+import { GoogleOAuthProvider } from '@react-oauth/google';
+
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <>
-    <GoogleOAuthProvider clientId={import.meta.env.VITE_GOOGLE_CLIENT_ID}>
-      <Provider store={store}>
-          <App />
-      </Provider>
-    </GoogleOAuthProvider>
-  </>
+  <GoogleOAuthProvider clientId={import.meta.env.VITE_GOOGLE_CLIENT_ID}>
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        <App />
+      </PersistGate>
+    </Provider>
+  </GoogleOAuthProvider>
 );
