+++ frontend/fivo-client/src/services/recommendUser.js
@@ -1,0 +1,6 @@
+import axiosInstance from "../utils/axiosInstance";
+
+export const fetchUserRecommendedStocks = async () => {
+  const response = await axiosInstance.get("/stocks/user/userRecommendations/");
+  return response.data;
+};
