+++ frontend/fivo-client/src/utils/axiosInstance.js
@@ -1,0 +1,19 @@
+import axios from "axios";
+
+const axiosInstance = axios.create({
+  baseURL: "http://localhost:8080/api",
+});
+
+// ✅ 요청 시 JWT 토큰 자동 삽입
+axiosInstance.interceptors.request.use(
+  (config) => {
+    const token = localStorage.getItem("accessToken");
+    if (token) {
+      config.headers.Authorization = `Bearer ${token}`;
+    }
+    return config;
+  },
+  (error) => Promise.reject(error)
+);
+
+export default axiosInstance;
